Squashed 'UnityProject/Assets/Zenject/Main/' changes from a209394..22b3719
22b3719 Changed to just use Fasterflect's Name() rather than our GetPrettyName() since it's better. Changed state machine class a bit and updated SceneCamera based on that.  Changed a few things in our installers (still WIP)
8722243 Added comment for last change
e831208 Changed to print a warning if any ITickables or IInitializables are added to the container but not bound to the interface.  Changed DerivesFrom method to be more explicit when given type is equal
f257c1a Added ability to get all concrete types from di container.  Also made Instantiator an instance type
68f7a52 Added test to multi-scene test scene to run all scenes and make sure there's no startup errors
b16c315 Renamed ToSingle to To when the specific instance is passed since it's not really a singleton in that case.  Added new version of ToSingle that is actually treated as a singleton
ee1b75c Tweaked some of the names in Zenject
5e56845 Added a unit test to verify the object graphs generated for all scenes (editor, player, freeplay, frontend, login).  This works in Nunit outside of unity since it doesn't need to instantiate anything (just validates that all dependencies can be found).  Refactored a few Zenject things to support this ability. Currently only validates the object graph created from IDependencyRoot so won't cover factories (though this can be added to the test easily).
3c5ec40 Added list of compatible platforms
dedf594 Updated Fasterflect DLL to be able to publish web player builds

git-subtree-dir: UnityProject/Assets/Zenject/Main
git-subtree-split: 22b37197aab5d8a67da17418b3de7ea840146d98